+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 - A Changelog! :)
 - The `search` subcommand has a new outformat `table`. Filter out fields beforehand with the `select()` function in Humio and pipe to `less -S` for readability if output is wide.
 - The `repo` subcommand now has an outformat option.
+- Subcommands that take a `--repo` option now accept view names.
+  - View names must match exactly, unless `--no-strict-views` is set.
+  - In addition these subcommands also take a `--ignore-repo` regex string with repository names to ignore. By default the suffixes `-qa` and `-test` are ignored. Pass the empty string to disable.
 
 ### Changed
 
-- The `repo` subcommand now lists views as well. The filter option now provides a regex pattern to filter out matching names. By default this option removes -qa and -test suffixes.
-- The `repo` subcommand's filter option has changed to `ignore`. Takes a regex-pattern matching repo names to ignore. By default all names ending with -test or -qa are ignored. Disabled if empty.
+- The `repo` subcommand now lists views
 - Turned sorting off by default to take advantage of streaming searches.
 - Bumped project dependencies.